+++ P_plats.c
@@ -238,7 +238,7 @@ int EV_DoPlat
     {
       case raiseToNearestAndChange:
         plat->speed = PLATSPEED/2;
-        sec->floorpic = sides[line->sidenum[0]].sector->floorpic;
+        sec->floorpic = sides[(unsigned short)(line->sidenum[0])].sector->floorpic;
         plat->high = P_FindNextHighestFloor(sec,sec->floorheight);
         plat->wait = 0;
         plat->status = up;
@@ -252,7 +252,7 @@ int EV_DoPlat
           
       case raiseAndChange:
         plat->speed = PLATSPEED/2;
-        sec->floorpic = sides[line->sidenum[0]].sector->floorpic;
+        sec->floorpic = sides[(unsigned short)(line->sidenum[0])].sector->floorpic;
         plat->high = sec->floorheight + amount*FRACUNIT;
         plat->wait = 0;
         plat->status = up;
